Fix x86 cDAC DBI stackwalk: Debugger_STRData layout, ParseContinuation overrun, and GetStackWalkCurrentContext ESP
On x86 the cDAC DBI path had three defects that broke ICorDebug attach and
async stackwalks when DOTNET_ENABLE_CDAC=1:

- Debugger_STRData.fp was a host-sized FramePointer (4 bytes on x86) while the
  managed cDAC treats it as a fixed-width value. Normalize fp to CORDB_ADDRESS
  so the native mscordbi consumer and the managed cDAC agree on struct layout;
  the RS converts to/from FramePointer at the boundary. ctx stays a host-sized
  pointer. Fixes the internal-frames (GetActiveInternalFrames) attach hang.

- ParseContinuation declared pDiagnosticIP as PCODE* (4 bytes on x86) while the
  managed cDAC writes 8 bytes, overrunning the caller's stack. Change the native
  signature to CORDB_ADDRESS* to match.

- GetStackWalkCurrentContext applied an x86 callee-popped-args adjustment that
  the native implementation does not: native builds the returned context from
  REGDISPLAY->pCurrentContext, and the M2U adjustment writes only REGDISPLAY->SP,
  not pCurrentContext->Esp. Return the raw context and remove the now-dead
  X86ESPIgnoresCalleePoppedArgs mechanism.

+++ src/native/managed/cdac/Microsoft.Diagnostics.DataContractReader.Legacy/Dbi/IDacDbiInterface.cs
@@ -195,7 +195,7 @@ public struct Debugger_JITFuncData
     public ulong vmNativeCodeMethodDescToken;
     public Interop.BOOL fIsFilterFrame;
     public ulong parentNativeOffset;
-    public ulong fpParentOrSelf;
+    public ulong fpParentOrSelf;                   // FramePointer (CORDB_ADDRESS)
     public Interop.BOOL isInstantiatedGeneric;
     public Interop.BOOL justAfterILThrow;
 }
@@ -224,14 +224,10 @@ public struct DebuggerIPCE_STRData_StubFrame
     public int frameType;                          // CorDebugInternalFrameType
 }
 
-// Holds data for each stack frame or chain. This data is passed from the RC to
-// the DI during a stack walk. Mirrors the native Debugger_STRData struct
-// defined in src/coreclr/debug/inc/dbgipcevents.h.
-//
-// `ctx` is a pointer into dbi-allocated memory.
-// The DAC writes the populated context through this pointer rather
-// than storing it inline. Code paths that do not produce a context
-// (e.g. EnumerateInternalFrames for cStubFrame entries) leave it as 0.
+// Holds data for each stack frame or chain, passed from the RC to the DI during a
+// stack walk. Mirrors the native Debugger_STRData in src/coreclr/debug/inc/dacdbistructures.h.
+// ctx is a host-sized pointer to a dbi-allocated DT_CONTEXT buffer the DAC writes through;
+// paths that produce no context (e.g. EnumerateInternalFrames cStubFrame entries) leave it 0.
 [StructLayout(LayoutKind.Explicit)]
 public struct Debugger_STRData
 {
@@ -242,10 +238,11 @@ public struct Debugger_STRData
         cRuntimeNativeFrame = 2,
     }
 
-    [FieldOffset(0)] public ulong fp;                           // FramePointer
-    [FieldOffset(8)] public ulong ctx;                          // DT_CONTEXT*
+    [FieldOffset(0)] public ulong fp;                           // FramePointer (CORDB_ADDRESS)
+    [FieldOffset(8)] public nuint ctx;                          // DT_CONTEXT* (host-sized)
     [FieldOffset(16)] public ulong vmCurrentAppDomainToken;     // VMPTR_AppDomain
     [FieldOffset(24)] public EType eType;
+    // v (method frame) and stubFrame overlap, mirroring the native anonymous union.
     [FieldOffset(32)] public DebuggerIPCE_STRData_MethodFrame v;
     [FieldOffset(32)] public DebuggerIPCE_STRData_StubFrame stubFrame;
 }
